requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-WTF==1.2.2
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.12.0
python-dotenv==1.1.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
